feat: logic extraction from history component
